requirements.txt
streamlit==1.25.0
langchain==0.2.0
openai==0.27.0
faiss-cpu==1.7.4
pandas==2.0.3
python-dotenv==1.0.0
